Improve flow expiration
Expiration buckets are defined in seconds and are unordered sets. When
the maximum number of flows is reached in the same second, the first
flow in the expiration bucket is deleted. But this flow does not always
correspond to the oldest flow and sometimes correspond to the last flow
being analysed. This can lead to issues when we need several packets to
analyse a flow content (case of SSL records on several packets for
example).
This commit improves the expiration deletion by adding a new timestamp
indicating last time the flow was inserted. This new timestamp is
computed in millisecond and is used when a flow of an expiration bucket
must be deleted. With this timestamp, we can determine the oldest flow
in the bucket (the oldest flow being inserted) and deleted this one.